Better parallelization
We are updating the way parallelization works in RHermes to be more standardized and easier to use. Instead of saving a BiocParallelParam object in the RHermesExp, the user can register() it at a global level so it's available to all RHermes functions.

This means setCluster() no longer returns an RHermesExp object but instead returns the most appropiate backend for the user. The function should be used jointly with register() like this: register(setCluster()).

Big thanks to @jorainer for suggesting this update :+1: